+++ src/repr.rs
@@ -9,7 +9,9 @@ use std::io;
 use {alloc::ffi::CString, alloc::format, alloc::vec::Vec, core::ffi::CStr};
 
 #[cfg(feature = "std")]
-use crate::{ValidationError, ValidationResult, WriteError, WriteResult};
+use crate::{WriteError, WriteResult};
+
+use crate::{ValidationError, ValidationResult};
 
 /// Validation of entities and other items
 pub trait CheckWritable {